Split MathLib
