+++ Source/ac-library-csharp/Math/MathLib.PowMod.cs
@@ -5,15 +5,9 @@ namespace AtCoder
 {
     public static partial class MathLib
     {
-        /// <summary>
-        /// <paramref name="x"/>^<paramref name="n"/> mod <paramref name="m"/> を返します。
-        /// </summary>
-        /// <remarks>
-        /// <para>制約: 0≤<paramref name="n"/>, 1≤<paramref name="m"/></para>
-        /// <para>計算量: O(log<paramref name="n"/>)</para>
-        /// </remarks>
+        /// <inheritdoc cref="ModCalc.PowMod(long, long, int)" />
         [MethodImpl(256)]
         public static long PowMod(long x, long n, int m)
-            => InternalMath.PowMod(x, n, m);
+            => ModCalc.PowMod(x, n, m);
     }
 }
